+++ src/frontend/src/lib/components/nfts/NftDisplayGuard.svelte
