+++ src/frontend/src/lib/components/nfts/NftHero.svelte
@@ -4,7 +4,7 @@
 	import NetworkLogo from '$lib/components/networks/NetworkLogo.svelte';
 	import NftActionButtons from '$lib/components/nfts/NftActionButtons.svelte';
 	import NftBadge from '$lib/components/nfts/NftBadge.svelte';
-	import NftImageConsent from '$lib/components/nfts/NftImageConsent.svelte';
+	import NftDisplayGuard from '$lib/components/nfts/NftDisplayGuard.svelte';
 	import NftMetadataList from '$lib/components/nfts/NftMetadataList.svelte';
 	import BgImg from '$lib/components/ui/BgImg.svelte';
 	import BreadcrumbNavigation from '$lib/components/ui/BreadcrumbNavigation.svelte';
@@ -48,15 +48,15 @@
 <div class="relative overflow-hidden rounded-xl" in:fade>
 	<div class="relative h-64 w-full overflow-hidden">
 		<div class="absolute h-full w-full">
-			<NftImageConsent {nft} showMessage={false} type="hero-banner">
+			<NftDisplayGuard {nft} showMessage={false} type="hero-banner">
 				<BgImg imageUrl={nft?.imageUrl} size="cover" styleClass=" blur" />
-			</NftImageConsent>
+			</NftDisplayGuard>
 		</div>
 
 		{#if nonNullish(nft?.imageUrl)}
 			<div class="absolute flex h-full w-full items-center justify-center text-center">
 				<div class="relative flex h-[90%] overflow-hidden rounded-xl border-2 border-off-white">
-					<NftImageConsent {nft} type="nft-display">
+					<NftDisplayGuard {nft} type="nft-display">
 						<button
 							class="block h-auto w-auto border-0"
 							onclick={() =>
@@ -67,7 +67,7 @@
 						>
 							<Img src={nft.imageUrl} styleClass="max-h-full max-w-full" />
 						</button>
-					</NftImageConsent>
+					</NftDisplayGuard>
 					<span class="absolute bottom-0 right-0 m-2.5">
 						<NetworkLogo color="white" network={nft.collection.network} size="xs" />
 					</span>
